small typedef fixes and folder structure fixes
