+++ server/schema/typeDefs.js
@@ -20,7 +20,7 @@ const typeDefs = gql`
 		description: String
 		index: Int
 		userId: String
-		boardId: String
+		boardId: ID
 		status: String
 	}
 
@@ -35,7 +35,7 @@ const typeDefs = gql`
 		profile(profileId: ID!): Profile
 		tasks: [Task]
 		taskBoards: [TaskBoard]
-		tasksById(boardId: String!): [Task]
+		tasksById(boardId: ID!): [Task]
 	}
 
 	type Mutation {
@@ -47,7 +47,7 @@ const typeDefs = gql`
 			title: String!
 			description: String
 			userId: String
-			boardId: String!
+			boardId: ID!
 		): Task
 		updateTaskStatus(_id: ID!, status: String!): Task
 	}
